Clarify the priority of ALPN and fix possible flaw
The previous text is ambiguous and can be read in two different ways:

Option 1: if the server recieves ALPNs "radius/1.0" and
"some-other-alpn-name" and you read "only one ALPN name which matches"
as "there must be only one ALPN, and if that one matches", then this
would force the server to abort the connection at this point.

Option 2: If you read "only one ALPN name which matches" as "there can
be multiple ALPNs, but only one matching the string" then we have the
problem of double-Spec if "radius/1.0" and "radius/1.1" was sent.

This new text makes the behavior explicit by first mentioning RADIUS/1.1
and then defining historic RADIUS/TLS as fall-back.